Implemented User's email, createdAt and updatedAt fields. Added more functions for filtering and searching info about users. Improved logic for API UserController

+++ src/main/kotlin/kz/mukha/fcs/controller/UserController.kt
@@ -2,7 +2,13 @@ package kz.mukha.fcs.controller
 
 import kz.mukha.fcs.dto.UserDto
 import kz.mukha.fcs.service.UserService
+import kz.mukha.fcs.util.LoggerUtil.getLogger
+import kz.mukha.fcs.util.ResponseCode
+import kz.mukha.fcs.util.ResponseUtil.createResponse
+import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
+import org.springframework.security.core.context.SecurityContextHolder
+import org.springframework.security.core.userdetails.UserDetails
 import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.PostMapping
@@ -17,15 +23,44 @@ class UserController(
     private val userService: UserService,
 ) {
 
+    private val logger = getLogger(this::class.java)
+
     @PostMapping("/register")
-    fun registerUser(@RequestBody userDto: UserDto): ResponseEntity<UserDto> {
-        val user = userService.registerUser(userDto.username, userDto.password)
-        return ResponseEntity.ok(UserDto(user.username, user.password))
+    fun registerUser(@RequestBody userDto: UserDto): ResponseEntity<Map<String, Any>> {
+        if (userService.userExistsByUsernameOrEmail(userDto.username, userDto.email)) {
+            logger.warn("Attempt to register with existing username or email: `${userDto.username}`, `${userDto.email}`")
+            return createResponse(
+                httpStatus = HttpStatus.CONFLICT,
+                responseCode = ResponseCode.USER_EXISTS,
+            )
+        }
+        val user = userService.registerUser(userDto.username, userDto.password, userDto.email)
+        logger.info("User `${user.username}` registered successfully with email ${user.email}")
+        return createResponse(
+            httpStatus = HttpStatus.OK,
+            responseCode = ResponseCode.USER_CREATED,
+        )
     }
 
     @DeleteMapping("/{id}")
-    fun deleteUser(@PathVariable id: UUID): ResponseEntity<String> {
-        userService.deleteUser(id)
-        return ResponseEntity.ok("User with id $id was deleted")
+    fun deleteUser(@PathVariable id: UUID): ResponseEntity<Map<String, Any>> {
+        val currentUser = SecurityContextHolder.getContext().authentication.principal as UserDetails
+        val username = currentUser.username
+        if (username != userService.findUserById(id).username) throw IllegalAccessException("Attempt to delete user with id `$id` from another user `$username`")
+
+        if (userService.userExistsById(id)) {
+            userService.deleteUser(id)
+            logger.info("User with id `$id` deleted successfully")
+            return createResponse(
+                httpStatus = HttpStatus.OK,
+                responseCode = ResponseCode.USER_DELETED
+            )
+        } else {
+            logger.warn("Attempt to delete non-existing user with id `$id`")
+            return createResponse(
+                httpStatus = HttpStatus.NOT_FOUND,
+                responseCode = ResponseCode.USER_NOT_FOUND
+            )
+        }
     }
 }
